Added regex for readProp. Needs to be checked. Added CMakelist extra includes. StartedToAdd Target reading in readProp. Now using iterators for vector in cDirectory. StartedToAdd aplhab sorting. Added detection for files like A.OS.h

+++ FileSystem/FileSystem.FileSystem.h
@@ -24,7 +24,7 @@ public:
     const std::string FavoritePath( unsigned int iPath ) const;
 
     /** How many directories are in favorites ? */
-    unsigned int FavoriteCount() const;
+    unsigned long FavoriteCount() const;
 //@}
 
 public:
